app/requirements.txt
blosc2==2.7.1
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.54.1
h5py==3.12.1
influxdb-client==1.46.0
kiwisolver==1.4.7
msgpack==1.1.0
ndindex==1.9.2
numexpr==2.10.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
planetary-coverage==1.1.1
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
radem @ git+https://github.com/AGHRADEM/radem.git@713f7a08aa414f65d0fc65c9dfdf0a7dfeee6399
reactivex==4.0.4
scipy==1.14.1
six==1.16.0
spacepy==0.6.0
spiceypy==6.0.0
tables==3.10.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3